[Improvement of responsiveness] Navbar + Image change
Made the Navbar collapse into a dropdown that links to the other pages as soon as a certain width is reached to make the website more fitting for mobile users. The set width for collapsing the nav bar also hides the image as that would be distracting for mobile users.